refactor(storage): separate interfaces from SQLite implementation
Split index.ts into types.ts (interfaces), sqlite-cache-store.ts, and
sqlite-project-store.ts. Public API unchanged.
